requirements.txt
requests
jinja2
python-dotenv
